+++ src/test/pythonEnvironments/info/interpreter.unit.test.ts
@@ -9,7 +9,7 @@ import {
 } from 'typemoq';
 import { StdErrError } from '../../../client/common/process/types';
 import { Architecture } from '../../../client/common/utils/platform';
-import { buildPythonExecInfo } from '../../../client/pythonEnvironments/exec';
+import { buildPythonExecInfo } from '../../../client/pythonEnvironments/base/exec';
 import { getInterpreterInfo } from '../../../client/pythonEnvironments/info/interpreter';
 import { EXTENSION_ROOT_DIR_FOR_TESTS } from '../../constants';
